+++ src/controllers/solidity/index.ts
@@ -168,14 +168,6 @@ export const soliditySetCommand: Command = {
 
         },
         {
-            name: "stdlib",
-            alias: "s",
-            title: "Stdlib version (version number or `latest`)",
-            type: "string",
-            defaultValue: "",
-
-        },
-        {
             name: "force",
             alias: "f",
             title: "Force reinstall even if up to date",
@@ -196,12 +188,10 @@ export const soliditySetCommand: Command = {
         } = {};
         if (args.compiler !== "") {
             versions.compiler = args.compiler;
+            versions.stdlib = args.compiler;
         }
         if (args.linker !== "") {
             versions.linker = args.linker;
-        }
-        if (args.stdlib !== "") {
-            versions.stdlib = args.stdlib;
         }
         await Component.setVersions(terminal, args.force, components, versions);
     },